+++ changelog.md
@@ -24,6 +24,7 @@
 - Fix shader banners and potion fluids using the wrong colors (BluSunrize)
 - Fix accumulator backpack getting rendered as full-bright when worn (BluSunrize)
 - Fix IE's tools rendering inverted in offhand (BluSunrize)
+- Fix crusher and excavator particles behaving weirdly (BluSunrize)
 
 ##### Version 1.20.4-11.3.0-174
 - Re-add CC-Tweaked integration (Malte)